+++ Shared/Common/Extensions/LoginRecord+.swift
@@ -13,3 +13,15 @@ extension LoginRecord: Equatable {
             rhs.password == rhs.password
     }
 }
+
+@available(iOS 12, *)
+extension LoginRecord {
+    open var passwordCredentialIdentity: ASPasswordCredentialIdentity {
+        let serviceIdentifier = ASCredentialServiceIdentifier(identifier: self.hostname, type: .URL)
+        return ASPasswordCredentialIdentity(serviceIdentifier: serviceIdentifier, user: self.username ?? "", recordIdentifier: self.id)
+    }
+    
+    open var passwordCredential: ASPasswordCredential {
+        return ASPasswordCredential(user: self.username ?? "", password: self.password)
+    }
+}
