--- Shared/Common/Extensions/NSData+.swift
@@ -1,11 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-import Foundation
-
-extension NSData {
-    func base64URLEncodedStringWithOptions(options: NSData.Base64EncodingOptions) -> String {
-        return ""
-    }
-}
