+++ lockbox-ios/View/ItemListView.swift
@@ -63,14 +63,14 @@ extension ItemListView: ItemListViewProtocol {
     }
 
     func bind(scrollAction: Driver<ScrollAction>) {
-//        scrollAction.delay(RxTimeInterval(0.1))
-//                    .drive(onNext: { action in
-//                        switch action {
-//                        case .toTop:
-//                            self.tableView.scrollToRow(at: IndexPath(row: 0, section: 0), at: .none, animated: true)
-//                        }
-//                    })
-//                    .disposed(by: self.disposeBag)
+        scrollAction.delay(RxTimeInterval(0.1))
+                    .drive(onNext: { action in
+                        switch action {
+                        case .toTop:
+                            self.tableView.scrollToRow(at: IndexPath(row: 0, section: 0), at: .none, animated: true)
+                        }
+                    })
+                    .disposed(by: self.disposeBag)
     }
 
     var tableViewScrollEnabled: AnyObserver<Bool> {
